refactor: remove old and new isbn code overlap
