+++ pyproject.toml
@@ -205,7 +205,6 @@ citations = "papis.explorers.citations:cli"
 crossref = "papis.explorers.crossref:cli"
 dblp = "papis.explorers.dblp:cli"
 isbn = "papis.explorers.isbn:cli"
-isbn_standalone = "papis.explorers.isbn_standalone:cli"
 json = "papis.explorers.json:cli"
 lib = "papis.explorers.library:cli"
 yaml = "papis.explorers.yaml:cli"
@@ -229,7 +228,6 @@ dblp = "papis.importer.dblp:DBLPImporter"
 doi = "papis.importer.doi:DOIImporter"
 folder = "papis.importer.library:FromFolderImporter"
 isbn = "papis.importer.isbn:ISBNImporter"
-isbn_standalone = "papis.importer.isbn_standalone:ISBNStandaloneImporter"
 lib = "papis.importer.library:FromLibraryImporter"
 pdf2arxivid = "papis.importer.arxiv:ArxivFromPDFImporter"
 pdf2doi = "papis.importer.doi:DOIFromPDFImporter"
